Add another spell point to specialists.
TODO the extra slot needs to be used for a matching specialist spell
TODO also make sure the Pick button/method does this aswell
These two are not that important.

TODO also write the specialist spells of greater level to the spellbook
     or will it be done on levelup? Or was it just Edwin (1/level iirc)?
Any knowledge on this?

It also seems they don't get the extra spell slot later on.

svn-id: r4880