ImageMgr: Simplify GetPalette to only return the actual image palette.
Previously, BMPImporter would do different stuff for 24bpp images, but that
code was moved to Interface, so we can drop it now. GetPalette now does nothing
and reports an error in a>8bpp case.

Note: MOSImporter has a palette, but we treat it as 32bpp image, so it doesn't
implement GetPalette.

Signed-off-by: Tom Prince <tom.prince@ualberta.net>