cmake: Fix behaviour wrt cmake policy CMP0012.
Old versions of cmake interpret
IF(Yes)
by dereferencing a variables 'Yes' (which is unlikely to be a true value). In
particular, if variable 'option' has the value 'Yes', then
IF(${option}"
will evaluate to false.

CMP0012 changes this behaviour (introduced in 2.8.0) but there is no need to
trigger it.