Fix drawing of floor area animations in IWD/HOW/IWD2
In these games an area animation height <= 0 is now treated as
background, meaning it will be drawn before actors and not covered by
walls. It being drawn before actors is verified in the original games,
but the cover behaviour is unverified.

This behaviour is covered by the new GF_IMPLICIT_AREAANIM_BACKGROUND flag.