move Schedule check inside ValidTarget
This assumes that unscheduled actors can never be valid targets. This
way we can get rid of a bunch of separate checks and fix some situations
in which such a check was missing (e.g. party members running on AI
scripts would attack and kill monsters that were not even present).