implement respawning at spawnpoints
While most spawnpoints in BG2 are triggered once only, BG1 has a
lot of spawns that reactivate over time. This change implements
(some of) that respawning. I think we don't reactivate spawns as
often as the original, yet.

The patch also changes spawngroup spawns to be all-or-nothing
except it makes sure that at least one creature gets spawned for
any spawn that is activated no matter the levels involved.