be less strict when initiating dialog
Requiring line of sight can easily break some of the game scripts
when the party is positioned in a way that keeps some members
invisible. I've seen it break both BG1 and BG2 in places.

It also means that if "the party" sees someone you'd like to talk
to but the party leader himself doesn't, clicking on the target
won't initiate dialog.