implement MultiPlayerSync
This is mainly to fix cutscenes of the form:

LeaveAreaLUA(...)
MultiPlaySync()
MoveViewPoint(...)

Since the map hasn't changed yet when MoveViewPoint is reached,
the command fails (or gets overwritten later on when the map does
change). As a result the player is often left looking at the party
which isn't even supposed to be present in the scene while sounds
imply interesting things going on in other parts of the area.

Adding a wait in place of the MultiPlayerSync command makes sure
that the area changes before trying to set the new view point.