treat IE_AVATARREMOVAL as being unscheduled instead of just invisible
This fixes cases where two actors share the same location on the map with
one of them having the avatar removed. If the removed actor was found
earlier in the list of actors it would be impossible to talk to or
otherwise interact with the present actor (e.g. IWD AR9101).