AUTHORS: made fairer and saved fuzzie some mail
NOTE: git commit has an --author switch, which can be used to change the
author of a commit. Add -s to mark it as reviewed. Use it for external
contributions outside of git.

There are two modes:
  git commit -s --author="A U Thor <author@example.com>"
  git commit -s --author=author

The first will add the author with the specified data, while the second
will first check our history for mentions of the "author" substring and if
found copy the contact details from there (shortcut).