Font: make note of 2 possible gotchas
Portrait status icon regions are 1px too small for their rows.
This brought to my attention that *if* the blitter (not just the last glyph) were to overrun the edge of the region, then we could mistakenly consider printing done in a case where a long word was wider than the region, but we still had more lines that would land within the region. This doesnt appear to be naturally occuring.