add a bit of a hack to allow BAM font to have uncompressed data
I’m doing this for ease of development, but Fuzzie says it would be better if we could use the BAMImporter directly (instead of an AnimationFactory)
I’m unsure of how this would work because I’m afraid that some frames are shared for multiple characters in some of the translation packs.
Additionally I’m not sure we can assume 255 frames per cycle (we use the first byte of a 16bit char to access the frame, and the second for the cycle)
I dont know how we can match frames to a glyph without exposing more of BAMImporter to the public and rewriting it a bit (probably fine, but I have too much on my plate ATM).