Font: remove the anchor and clip arguments
it is not the fonts responsibility to do any clipping or positioning beyond using the text region
Add FIXMEs for the two special cases that needed the arguments (tooltips and overhead text). They can be easily fixed in the future.
We now have a much simpler interface with clear use and responsibility