TextSpan: tweaks to ivar semeantics
allow NULL palette and take from the font if NULL. the text buffer should only hold the string that is actually rendered, not the entire string passed.
also render the span at construct time. spans are meaningless if they arent rendered so we save nothing by using lazy instantiation of the sprite.