Region: fix off by one error in Intersect test
basis for change:
drawing a Region(0,0,10,10) and testing intersection with Region(10,0,10,10) would yield true even tho there is no overlap in the blitted rectangles. Region(11,0, 10, 10) was false as expected, but with a 1px gap between rects on screen. We should be able to have neighboring rects that "touch" without intersecting.

If this logic now seems odd, then we need to adjust the drawing code instead