Font: fix vertical/horizontal alignments (both multi and single line)
we never allowed some vertical alignments on multitext because it was impossible to know ahead of time how lines would break.
to compensate for this we can use our new RenderAsSprite function to render the text to a sprite then vertically align the sprite.
This probably wont ever be used because potential clients *should* be using a text container that already renders text to a sprite, but I want to be through.

Some functionality that is used is horizontal alignment for multiline text.
For this we render each line then blit the line to the screen/canvas, positioned acording to its alignment.

One last thing this fixes: our previous thoughts about single line printing were incorrect (by far).
All line calculations should be based off the maxHeight and not the bounding box for the line.
the primary reason this didnt work previously is because we didnt take line descent into consideration.
a second reason was our forced padding of 5px to all font printing, this mostly only affected the horizontal plane.

I have a comparison overlay of BG2 and GemRB with a 100% alignment match using these new calculations so Im quite confidant in them :)