TextContainer: provide more context to the internal drawing functions
we need both a screen frame and a point inside that frame.
this fixes the text not wrapping (such as with drop caps)

it should also be a bit faster since we can intersect with the drawing frame before passing to font->Print so that function wont have to uselessly iterate text that is cut from the region.