TextArea: more succinct method of determining dialog text position
The main reason I'm switching to this location method, is so that journal updates/XP rewards etc arent scrolled out (even if multiline!).

it has 1 potential drawback compared to the other method. It will show the response to any selections above the next text; I dont mind this, however, I understand if we want to revert or tweak the dialog mechanisim to reorder tings so that we can have journal updates without also seeing dialog response.