TextContainer: optimize performance by separating layout from drawing
The change isnt as big as it looks. Really its mostly moving parts of the draw methods into layout methods.

the OpenGL driver has to redraw everything *every* frame. that makes text layout+drawing really taxing. This compounds the problem with the GL driver not being optimized for font printing yet.

this has additional benifits aside from OpenGL tho. Now we can know content heights as soon as it is appended to a container, which makes dialog handling easier. We can use a cleaner TextArea::Draw method now as well.