Use find_package instead of including the FindPythonLibs file which does not work.
The cmake version is anyway restricted by the cmake_required directive. This
was leading to an incorrect version detection when using cmake 3.0 (MAJOR > 1
but MINOR < 7) and when Python3 libraries are available.

FindPythonLibs.cmake expects the PythonLibs_FIND_VERSION variable to be set by
find_package function in order to choose the correct version of python libraries.