don't call ProcessActions twice for area scripts
Without this change, ProcessActions would be called twice per tick for
area scripts. Among other problems that means that short blocking
actions, like for example SmallWait(1), are (sometimes) essentially
turned into non-blocking actions.

This is easily seen in the final scene in ToB after defeating Melissan
when all of the party members are supposed to initiate dialog. Without
this change some of the dialogs are skipped. Other scripts may break in
similar ways.