Revert "redraw inventory window when hitpoints change"
This reverts commit 4732945c197f2c012776b3f63f9a03c7853650f0.

unfortunately, this completely breaks all inventory subwindows :(

Easily observed by accessing the special abilities window or scrolling and item description for a sufficient amount of time (a few seconds). The main inventory window below the "modal" window begins receiving the events. Not sure what the best fix is, but it does seem odd that events can even be delivered to another window while one is modal, so fixing that behavior may be an option.

We will revisit this later, post release.