Label: try to hack around bad data causing PST store labels to be clipped
the labels for PST stores are 25px high, but the font is 13px per line causing the text to print in the wrong position (because it cant fit both lines).
this in turn causes the label to be clipped.

by forcing IE_FONT_SINGLE_LINE for labels with height <= font->LineHeight we can add a check in the drawing method to increase the print region to be 2 lines for multi line labels that dont have enough height for 2 lines by default.