pst: added spwi506 (ice and fire) override
this spell has two main projectiles for all intents and purposes, affecting
the same aoe. We achieve that by using the normal projectile and it's success
spell for the secondary fire damage and a cloned version for the ice one, that
we shoot with a delay from the cfb (or it would apply to all targets of
the primary projectile, creating multiple comets).

MEH!