fx_barbarian_rage: added basic fatigue malus
i think in the original you could negate it with the remove fatigue
spell, but that only resets real fatigue and fatigmod.2da is the
same as in bg2. So while we could add enough real fatigue to get the
icon in, the spell would still only remove fatigue and not the two
maluses too.