Interface: adds distinct encoding for file system path
Consider the case: IWD2, german (ISO-8859-1), located on a mb file
system (UTF-8). For reasons such as resource listing (sound files), we
shouldn't couple the TLK encoding when file paths are subject to
something different.

While this is a quite static setting for now, it's better to prepare
decoupling anyway.