Refactor Scrollbar
needed to rewrite now that its position is based on Control::Value
things are a little shorter and more straightforward now
the old method of calculating the nib position incured cumulative rounding errors.
We no longer need to keeep a separate varialbe for the nib position. Instead we calculate it at draw time as a simple % based on Value relative to the min/max.