OpenLevelUpWindow: Fix Sorcerer LevelUp (#57)
* OpenLevelUpWindow: Fix Sorcerer LevelUp

One of these checks is about spell selection. If the character class
has a MageTable, then a loop counts the difference in spell count
for each spell level and if there are new spells known, the spell
selection window is opened.

The problem is that for sorcerers it checks the MXSPLSRC table
which is how many spells the character can cast and not the proper
table which is SPLSRCKN and defines how many spells the character knows.

So, when a sorcerer levels up to a level that doesn't grant additional
known spells but grants more spell castings per day, the spell selection
window is opened but there are no spells to be selected and the leveling
process is stuck.