bg2::cg: update total ability points during recall
Update the total score to reflect the correct sum when the player
recalls a previously stored dice throw.

Commit 2f3ada03e25732ff added the displaying of total score
of ability points when rolling dice. However, this code is only run
during the RollPress function so when the player clicks
the Recall button to recall a former dice throw, this display
isn't updated to reflect this and still shows the previously
rolled result.

Signed-off-by: Jaka Kranjc <lynxlupodian@users.sourceforge.net>