View: small refactor to event dispatch
all events are routed publicly though the View base and an implementor method is called that returns bool for if the event was handled. unhandled events are bubbled up to the superview recursively.

We've added an "eventProxy" member that can be used to forward all events directly to the proxy view. We will use this mostly for windows -> scrollbars, but it could be useful for some of the other nested view composites (TextArea, ScrollView).

fixes issues with the view debugging code since we can now see all events in the base
fixes having to use the IgnoreEvents flag for some cases. we now can return false from the implementors methods.

issue #78