CreateCreatureCore: use LaunchPoint for traps instead of Pos
fixes trap spawns spawning enemies right on top of you, instead of at the
desired point. Very obvious in Firkraag's dungeon (ar1202) with the trap
right before Samia — the wolfweres should be summoned in the corridor just
before the elemental well. Same for golems and vampires. Visible also in
beholder lairs

CreateCreatureAtFeet is not used by default, so CC_OFFSET is fine for now