SDLVideo: split sprite rendering for SDL12 and SDL2
attempting to make the high level functions generic and implementing the details for each
mostly involves moving code from the base to SDL12 and implementing missing bits for SDL2

this makes the SDL2 driver functional, but some things remain incomplete (event handling, some effects, etc)
it should be noted that combining blitters means it is possible to attempt blitting of any sprite with any effect, but in SDL12 this wont always work correctly because we dont handle different pixel formats well
however, this should be limited to combinations that aren't actually used in practice. something to be aware of.