SpriteCover: use 50% alpha instead of dithering
I'll open an issue to discuss, but the dithering doesn't work well with SDL 2.
Sometimes it seems to result in banding where the "screen" doesn't quite align with the actual pixels

If we want we can limit this to SDL2 driver by having the SDL1 driver by moving the pixel manipulation to the video driver (using the same code in DrawPolygon) and adding a "Dither" function that simply fills with 50% alpha on SDL2.

However, it seems that BGEE also uses 50% alpha instead of dithering. My guess is the dithering was done to make things fast and use less memory. GemRB is using a whole byte per pixel on the SpriteCover regardless so we don't save any memory by retaining dithering and i doubt the performance impact is meaningful.

So unless we want to keep it for nostalgia sake, or add an option to switch between the 2 methods I'm leaving this as-is