HotKeys: address more FIXMES
Window: add support for unregistering callbacks

EventMgr: add support for multiple registrations for the same key
this is so that we can support using the same hot key for multiple things (unused currently)
the idea is that we keep a "stack" such that the most recent registration gets the event, but if it unregisters (such as on window close)
then the next in line will still be able to get the event. Normally we accomplish a similar behavior by registering with the Window, but this will only work for Controls and it is dictated by Window order, rather than registration order.

This will fix the problem we were having where we register a button with the esc key but that button opens a new window which also results in an esc key registration and things dont work as expected because only one button could have the esc key at a time