Actor: fix fatigue calculations
Fatigue used to be generated dynamically, using TicksLastRested as base.
This caused most fatigue-modifying effects to be overwritten on the next
tick.  Now, fatigue stat is preserved and respected.

Also, Modified[] stats are used more extensively; it is allowed since
fatigue code runs at the same time as other modified stats are updated.

Not addressed: Constitution bonus is still applied dynamically, but
testing in BG2 suggests this is not the original game behavior.