Actor::OverrideActions: tweak the way charmed pcs work
more info from the original:
It's part of "CGameSprite::SpecialStateProcessing()"

If the creature is
    STATE_CHARMED
    Base EA <= GOODCUTOFF
    Current EA == ENEMY

then it will:
AttackReevaluate([GOODCUTOFF],10)

That also clarifies the distinction of EA 254[EVILBYCHARM/CHARMED_PC] - it is not issued this attack action.
Which is why "Normal" Charm doesn't cause allies to attack the party, while "Dire" Charm does.