Fix up copy and paste support
if we just create a text event with the clipboard contents this will automatically work with anything you can type text into
I did notice a bug where the text pastes the first character on the wrong side of the cursor, however. Other than that it seems to work and paste text appropriately.