Window: implement IsDragable()
I want to be able to drag windows to other positions on screen with a key combo even if they don't have the Dragable flag set
I also want this to apply even if the have the IgnoreEvents flag set (message_win and company).

now if you hold the control key and drag on the window (not on one of its controls) you can move it.