Viewport: don't keep trying to scroll beyond map boundaries
if you use the map control to move the viewport to the edge of the map it will set the goal point
well beyond what GameControl allows. this will effectively lock your viewport in place.

we need to return a bool from GameControl so the GlobalTimer can know that it has scrolled to its maximum and reset the goal