Button: fix overlays and implement animated ones
issue #249

we had a FIXME/TODO about not being able to animate the coloring for the overlays
subviews gives us the ability to update things like this via WillDraw()

our colors in GUICommon.py were wrong too. I tested this using a specially crafted custom portrait in BGEE and found they are just doing a color multiply with almost pure red (no alpha)
I didn't bother to sample all the red colors, so the initial damage one may be off a bit, but the final color is correct.
using an alpha on a BLIT_TINTED actually modifies the alpha of the sprite you are blitting anyway, so its almost certainly not what you want to use except for some effects.