GUIClasses: CreateSubview should cleanup its old reference
when you call view.CreateSubview the following process happens:
1. a new view is created and its scripting reference exists in a generic "control" group bucket because we don't know what window it belongs to yet.
2. it is then inserted into a view hierarchy at which point it (most likely) will have a window and a new reference is created for the window group

this works fine most of the time. However, if we use an id for multiple controls and their lifetimes overlap this becomes a problem.
since we don't actually need the generic control group reference once the view is added to a window we can just remove that reference so that we can recycle the same id (as long as the controls belong to different windows)

should theoretically fix issue #395