GUIScript: fix some implementation oversights in AddSubview/RemoveView
in order to readd a previously removed view you need to be able to pass an id to AddSubview (we assume you aren't moving between windows still)
we were also always using GetScriptingRef() to construct the GSObject, but we should be using the ref we already decoded instead
we should also only call RemoveFromSuperview if we are deleting the view. this way we can remove references without the view disappearing.

ReparentSubview needs to pass the id of the view being replaced