Sprite2D: switch from desperate ints to a Region
this was long overdue. It has grown tiresome to have to construct Regions with these.
This allows us to use things like intersection directly on the sprites

this is mostly a find and replace except a few locations where we could obviously just use the new Region directly
the next commit will change the Video API to take Regions for constructing sprites to make the transformation complete