Torch SpriteCover
part of #441

in favor of using viewport stencils
it is realistically the best way to achieve sprite masking with SDL2
it replaces dozens of small textures (that are constantly getting created and destroyed) that might overlap and blend in funky ways with only 2 screen sized stencil buffers

unfortunately, in its current state, this is a performance hit. We need to address some of the FIXMEs/TODOs

Rendering on SDL 1.2 does not work correctly ATM (probably doesn't even build). We need to implement BLIT_BLENDED, BLIT_STENCIL_ALPHA, and BLIT_STENCIL_RGB to fix it.