Map: better check for BehindWall()
fixes #607
however, there probably still exists some ugly corner case someplace where an object can be behind one polygon and infront of another
we will need to handle that by detecting this case and then generating an object sized stencil for the offender.

this requires:
1. probably an Actor method to get a bounding box for the Actor + attached effects so we can properly test for the condition
2. stencil buffers need to somehow be given coordinates instead of being always considered screen coordinates

I'll add this to the #441 checklist.