Actor: remove BLIT_TRANSSHADOW
BLIT_TRANSSHADOW relies on the _tinter_ to actually work correctly (SDL 1.2) this is why it is broken in SDL 2. You can replicate the SDL 2 breakage by commenting out the tinter call or modifying it to select the "PALALPHA=true" template.

I realize that copying the palette and iterating it is not ideal for performance, but I also think we have lower hanging fruit there and this will also open the door for removing both BLIT_TRANSSHADOW and BLIT_NOSHADOW from the codebase entirely which should simplify a lot of the rendering pathways and recoup some of the palette copy performance hits.

we could also cache the copy to effectively eliminate the penalty but that is more complicated work for a later time (if we need it).

fixes #397