Actor: refactor DrawActorSprite
mostly just need a good std::vector for debugging purposes, but also realized there was no need for some of the parameters
if we handle them in Actor::Draw. this would be more efficient if we refactored CharAnimations to not use C style arrays, but that is a later thing.

added a TODO for having CharAnimations use some kind of std container and do the sorting itself. Actor shouldn't be responsible for the z-ordering I don't think