Actor: fix use-after-free
equipment changes alter the stance animations
when we get to UpdateActorState() these won't be around anymore
ideally, changing the animations would happen all at once in either UpdateActorState() or UpdateDrawingState(), but that is a bigger hassle than it is worth right now. This hack seems to work ok for now.