Improve formation behaviour
Currently, right-clicking to move the party in formation always orients the
formation in the same way.  This is different from Infinity Engine behaviour,
and inconvenient, because you have to rotate it into the desired orientation
every time you want to move.

With this change, the formation is assigned a base angle, which is computed
from the first active character's position and the clicked point, so that
when moving in one direction, the formation always stays pointed the same
way.  Rotations from moving the mouse now have a small dead zone applied to
them to prevent poor behaviour near the clicked point.