Fix SDL2 mouse wheel behaviour
Observed behaviour: regardless of which direction the wheel is moved,
a scrollbar always scrolls to the top.
With this change, we now use the same method as in the SDL 1.2 code,
using SetPos to adjust by ScrollDelta.  This appears to work nicely on
Linux with SDL2, but is untested elsewhere.