Allow the mouse wheel to work immediately in the Load screen
Using the mousewheel to scroll should immediately have an effect after
opening the Load screen, but currently you have to click the scrollbar
first.  The scrollbar needs to be focused.  Confusingly, there are two
types of focus, Focus and MouseFocus, with no clear indication why.
This change may not be entirely correct, but it seems to work - and
maybe focus handling should be changed entirely to have only one type
of focus.