Improve formation movement and right-click moves in general
This fixes a number of movement related issues by changing behaviour to
match the Infinity Engine more closely.

When a formation is initially created, the angle is computed from the
position of the click and the position of the first selected character.
Changing the orientation with mouse moves is still possible, as before,
but with a small deadzone to prevent out-of-control movement near the
position of the initial click.

Also, a single right-click now initiates a move, for a single character
or several of them, as in the Infinity Engine.  I believe this fixes
issue #629.

A note on the maths: I think the changed RotatePoint now follows the
usual mathematical definition, which allows for a simpler
AngleFromPoints.  The only thing to keep in mind is that the coordinate
system has the Y axis increasing downwards, but since everything is
consistent now, it all works out.