Unify SDL keycode translation somewhat
The motivation for this was the handling of tab - it was reported as
GEM_TAB on key press, but just plain 9 on release.  Now both paths
use the same translation function, although there are still some
differences regarding modifiers.

Beyond the things it fixes, this should be equivalent in behaviour to
the previous code, since key is initialized from sym (so the test
for sym < 256 in the KeyUp case was meaningless).