Fix autopause on enemy sighted
There were a number of problems here.  The first one is that directly
after creating a character, BG2 goes into pause mode, which looks like
a lockup to the user, since the GUI isn't shown at that point.  The
issue is that Actor::UpdateActorState doesn't check for visibility
and triggers autopause for the Cambion, which is somewhere else
entirely.  I believe this code isn't needed at all, and deleted it.

The next problem is the check in Map::InitActor.  This has an
isVisible check using actor->Pos, but the position isn't set at that
point, it's 0,0.  Most callers of AddActor appear to follow up with
SetPosition (and you can't swap the two).  I removed the
ActorSpottedByPlayer call in this function, and added a log message
to show an error in case the condition ever triggers, which really
ought not to be possible.

The only semi-reasonable autopause code was in GenerateQueues, and now
we generate all enemy-sighted pauses from there.  It was necessary to
separate the autopause code from ActorSpottedByPlayer, so that we can
also call it for already IF_ACTIVE actors.

The final issue was the use of AnyPCInCombat.  That seems to just be
a countdown, which was observable in Chateau Irenicus.  After
defeating the Ogre Mage spawned in by the genie, and then walking to
the first group of kobolds.  If you do it quickly, AnyPCInCombat
returns true, and no autopause is generated, which is incorrect.
Instead of this, we now keep track of whether hostile actors are
visible on the map, and suppress new autopause events when we already
saw some.  That is bug #630.