Eliminate FreePalette calls
Most of the time, this can be removed entirely (when the PaletteHolder
is about to be destroyed), or replaced with an assignment of nullptr.
Also, PaletteHolders don't need to be initialized to null in
constructors.

There is also one fix for an issue in BAMImporter, where we could
potentially free memory without zeroing out the pointer, and then
return early.

There are remaining uses in CharAnimations, which are much more
complicated and are left for later.