Convert the ResourceSource plugins to be built-in
gdb can be very slow to start gemrb; it is probably reevaluating
breakpoints after every load of a plugin.
This moves all the ResourceSource code into core.  This simplifies the
code, and reduces the amount of plugins to be processed.  There's no
reason for them to be plugins, since it's not like the program can
function without them: Interface.cpp has hard dependencies on all of
them.