InventoryCommon: Allow DisplayItem to be used to cycle items
The way this was done before did not seem to work anymore, it was for some reason not refreshing the info window contents properly after closing and reopening it.

This method of cycling avoids closing the window, by allowing it to be refreshed instead.

The slots to be cycled are also extended to all possible slots, since the original allows you to inspect everything, even your fist weapon, for information.