Don't create a drag op if DragDropCreate is unsupported
This seems like just an oversight. We currently create drag ops even for
buttons that don't support them, which was one of the causes for the
weird cursor behaviour previously.

On the other hand, this made portrait drag work by accident - portraits
do not have drag/drop handlers associated with them, so this also
provides an override in Button to deal with that case.