pst/GUIREC: Actually calculate TNO's gained prof points
The Nameless "specific" check error was preventing that block from ever running.

This moves the proficiency point calculation up a level so it can be used by TNO

This also accounts for the extra 4 points that TNO gets compared to the other NPC's, which are cheated out of their starting points.