VisualFX: add a separate field for game position
now effects have a position in map coordinates and an offset vector (eg for casting animations)
it should be possible now to "rotate" an effect around a point without affecting the origin. for example, the visual effect for casting spells should move as the actor rotates to face the target.

as a bonus, this simplifies state updates/drawing

needed to finish #1119