Fixed segmentation failure when screenshot during video playback.
How to recreate: make a new character in BG1 campaign, skip the intro,
and after spawning on the map quickly go and save manually. The game
will freeze, or crash.
The GetScreenshot() function takes a VideoBufferPtr reference as a
second argument and only uses it in the if-else statement, while cast
was made for drawingBuffer pointer, which can (and will) be nullptr if
the audio from the intro (or any other cinematic) is still playing.