AREImporter: actually calculate the tile props size
instead of assuming the images are all the correct size
still not sure of the precise fix we need. It looks like sometimes one or more of the "images" are cropped off.
My best guess is that we need to layer them at the origin, which we previously assumed was top left.
The only example area I have is BG2 AR0235 which doesn't really answer the question as the image is cropped only on the horizontal

#1436