Interface: make GetProjectileServer noexcept
gcc is still complaining about Projectile not being able to be no except
I've looked at all the members and made sure they are noexcept constructible (did i miss one?)

I see the ctor is calling something that isn't noexcept so I'm hoping that is the cause. if not, somebody with gcc will have to look at it.