GameScript: unionize string parameters with ResRef and ieVariable
this should be much more efficient as we eliminate a bunch of temporary copies
we also eliminate a ton of implicit conversions

there is now added safety in general due to parameter mismatches, however, one scenario that could bite us is
writing to a parameter as a longer type (ieVariable) then using it as a shorter type (ResRef).
previously this would just truncate to 8 chars with a null terminator. if we then get the CString() for the ResRef it wont be null terminated at 8 chars
it will instead null terminate somewhere between the 9th and 64th char. At least it wont crash :)

something to keep an eye out for. I'm of the opinion that this was a bug since it doesn't make sense to assign as a larger type then use as a shorter one