FindSDL2.cmake: try to search harder for the headers
When searching for frameworks, if the file is specified as A/b.h,
then the framework search will look for A.framework/Headers/b.h.
If that is found the path will be set to the path to the framework.
CMake will convert this to the correct -F option to include the file.