MoviePlayer::Stop: added workaround for occasional fade problems
when you finish the iwd prologue a cutscene starts. It shows a video,
teleports everyone to the next area and there a hermit approaches you for
dialog. Sometime we end up showing a completely black screen, breaking
the game — only the exit window can be drawn. It seems that faking the
continuation of the automatic fade reset avoids this problem.