MapReverb: refactor to fix bad initialization
#1402

MapReverb previously relied on Map fields to initialize itself properly, but those fields don't get set at construction time.

After analyzing, it seemed like the best thing to do is to break the dependency MapReverb has on Map so we can construct the MapReverb when assigning the ambients to the Map.
We will internalize MapReverb into Map since the audio driver only requires MapReverbProperties. Instead of constructing MapReverb with the Map, we have 2 ctors: either pass a reverbID directly from the ARE, or fallback to the reverbs table and match on the WEDResRef.