Map::CanFree: properly check for cutscene runner, amend 1756ff60c
fixes saemon not starting his dialog on the attacked ship, because the one
from the previous area was still there. Also addressed by a fixpack 11 fix,
but the way we do this now is more correct anyway.

... on the other hand, gaelan would be broken again, so add a workaround. While
the scripted waits look sufficient, due to staggering of script updates (1/s),
the StartDialogNoSet in cut42a runs sooner than he is teleported in