SDLSurfaceSprite: revert changes to copy ctor
I thought it was relevant for #1606 since the bisect blamed e91dee1, but i'm now of the opinion that actually fixed a bug that was hiding another.

we shoudln't need to copy the render cache, since it will get created (with potentially different flags) when the sprite is rendered to screen.

regardless, i'm choosing to be conservative with sprite changes until after the release, and at some point it would be nice to dispense with the render cache anyway.
The future lies with shader accessible palettes.