EffectQueue: fix decrementing level immunity
These kinds of protective spells (incl. bounce and spell trap) include a
separate effect for each affected spell level, but as the protection
depletes, all of these effects should be updated at the same time.  The
new DecreaseParam1OfEffect() does just that.  It also now returns false
if the effects are already spent, but not yet removed -- this lag used
to sometimes allow them to absorb more levels than intended.