EffectQueue: remove selfcast check from opcode imm
Case in point: Durlag's Goblet, which lowers morale on use.  In IE, it
can be made safe with protection from fear, but with that check, this
doesn't work.  Note that the Goblet's ability is not flagged as hostile,
so it's not possible to distinguish it from benign effects in such way.